Split Sentry logic between UI and main.
